Weave and tangle at once, complete argument support.
Literati now has an 'all' action that will both tangle and weave the
files you give it.
You can also pass in a path to an individual file, an entire directory
to be operated on, and a mix of the above (multiple files and a
directory) and literati will process them all for you dutifully.